cdk-fargate/requirements.txt
aws-cdk.cdk
aws-cdk.aws-ec2
aws-cdk.aws-ecs
aws-cdk.aws-ecs-patterns
aws-cdk.aws-ssm